refactor: :truck: clean up folder structure and naming
