+++ src/renderer/views/AddProject/subcomponents/ProjectDetailsInput/style.ts
